+++ saleor/graphql/checkout/mutations/checkout_shipping_method_update.py
@@ -6,7 +6,6 @@ from ....checkout.error_codes import CheckoutErrorCode
 from ....checkout.fetch import fetch_checkout_info, fetch_checkout_lines
 from ....checkout.utils import (
     delete_external_shipping_id,
-    get_checkout_metadata,
     invalidate_checkout,
     is_shipping_required,
     set_external_shipping_id,
@@ -209,7 +208,6 @@ class CheckoutShippingMethodUpdate(BaseMutation):
             checkout_info, lines, delivery_method=delivery_method
         )
 
-        delete_external_shipping_id(checkout=checkout)
         checkout.shipping_method = shipping_method
         invalidate_prices_updated_fields = invalidate_checkout(
             checkout_info, lines, manager, save=False
@@ -220,7 +218,8 @@ class CheckoutShippingMethodUpdate(BaseMutation):
             ]
             + invalidate_prices_updated_fields
         )
-        get_checkout_metadata(checkout).save()
+        if delete_external_shipping_id(checkout=checkout):
+            checkout.metadata_storage.save()
 
         call_checkout_info_event(
             manager,
@@ -261,7 +260,7 @@ class CheckoutShippingMethodUpdate(BaseMutation):
             ]
             + invalidate_prices_updated_fields
         )
-        get_checkout_metadata(checkout).save()
+        checkout.metadata_storage.save()
         call_checkout_info_event(
             manager,
             event_name=WebhookEventAsyncType.CHECKOUT_UPDATED,
@@ -274,7 +273,6 @@ class CheckoutShippingMethodUpdate(BaseMutation):
     @classmethod
     def remove_shipping_method(cls, checkout, checkout_info, lines, manager):
         checkout.shipping_method = None
-        delete_external_shipping_id(checkout=checkout)
         invalidate_prices_updated_fields = invalidate_checkout(
             checkout_info, lines, manager, save=False
         )
@@ -284,7 +282,8 @@ class CheckoutShippingMethodUpdate(BaseMutation):
             ]
             + invalidate_prices_updated_fields
         )
-        get_checkout_metadata(checkout).save()
+        if delete_external_shipping_id(checkout=checkout):
+            checkout.metadata_storage.save()
 
         call_checkout_info_event(
             manager,